refactor: remove ETH authentication (#20)
Because ETH authentication was only used to authenticate requests to Highlight
(and not authentication payload) it's usefulness is limited as this
cannot be used to validate user requests (unless users interact with
Highlight directly through their wallet, by adding dummy network
communicating with Highlight RPC).

We are going away from this approach, where there will be different method (if any)
of authenticating access to `postJoint`, for the time being we will focus
on authenciating messages within agents (for example validating payload
using provided signature).

This PR caused following changes:
1. `hl_postJoint` is now entry point for submitting units. It's not authenticated.
2. Unit has simplified format, there is no hash (because we don't have unique information
   to avoid duplicates - nonce has been removed), txData is now inlined and only includes
   toAddress and data. There is no signature either.