refactor: index units instead of events
This change is needed for indexer to know unit timestamp when indexing
(so alias timestamp can be taken from unit instead of user provided).

Currently Highlight indexes events as main chain (event IDs are used
as MCI). They are also stored redundantly (as part of events and
unit_events).

Because events are only loosely linked to unit, and by default stored
with no unit information at all it's impossible to retrieve unit when
indexing by event.

This PR changes underyling architecture of Highlight to make units the
main object in the chain, that events belong to, this means:
1. We no longer duplicate events in storage.
2. MCI is now refering to units (that contain events) instead of event
   counters.

As part of it, we remove old `getEvents` method that was using event
IDs.

With this change Checkpoint can now index Highlight unit-by-unit
having access to unit (and its timestamp).

Step counting has been also updated as it seemed a bit confusing, not
sure I understand original intention (multiple events counted as only
single step, and once they were saved it added another step). That means
that if we had 10 events emitted it will be 2 steps.

I updated this logic so each event emitted counts as 1 step.